requirements.txt
Appium-Python-Client==5.2.4
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
diskcache==5.6.3
filelock==3.19.1
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.10
huggingface-hub==0.35.3
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
joblib==1.5.2
lazy_loader==0.4
llama_cpp_python==0.3.16
MarkupSafe==3.0.3
networkx==3.5
numpy==2.2.6
opencv-contrib-python==4.12.0.88
outcome==1.3.0.post0
packaging==25.0
pillow==11.3.0
PySocks==1.7.1
PyYAML==6.0.2
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
selenium==4.35.0
setuptools==80.9.0
sniffio==1.3.1
sortedcontainers==2.4.0
threadpoolctl==3.6.0
tifffile==2025.9.20
tokenizers==0.22.1
tqdm==4.67.1
transformers==4.57.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
urllib3==2.5.0
websocket-client==1.8.0
wheel==0.45.1
wsproto==1.2.0
